+++ api4/system.go
@@ -21,6 +21,7 @@ import (
 	"github.com/mattermost/mattermost-server/v6/model"
 	"github.com/mattermost/mattermost-server/v6/services/cache"
 	"github.com/mattermost/mattermost-server/v6/services/upgrader"
+	"github.com/mattermost/mattermost-server/v6/shared/eventbus"
 	"github.com/mattermost/mattermost-server/v6/shared/mlog"
 )
 
@@ -71,6 +72,13 @@ func (api *API) InitSystem() {
 	api.BaseRoutes.System.Handle("/onboarding/complete", api.APISessionRequired(getOnboarding)).Methods("GET")
 	api.BaseRoutes.System.Handle("/onboarding/complete", api.APISessionRequired(completeOnboarding)).Methods("POST")
 	api.BaseRoutes.System.Handle("/schema/version", api.APISessionRequired(getAppliedSchemaMigrations)).Methods("GET")
+
+	// GET /api/v4/events/topics
+	api.BaseRoutes.System.Handle("/events/topics", api.APISessionRequired(getEventsTopics)).Methods("GET")
+	// GET /api/v4/events/topics/<topicId>
+	api.BaseRoutes.System.Handle("/events/topics/{topic_id}", api.APISessionRequired(getEventsTopic)).Methods("GET")
+	// GET /api/v4/events/schemas/<topicId>
+	api.BaseRoutes.System.Handle("/events/schemas/{topic_id}", api.APISessionRequired(getEventsSchema)).Methods("GET")
 }
 
 func generateSupportPacket(c *Context, w http.ResponseWriter, r *http.Request) {
@@ -1007,4 +1015,76 @@ func checkHasNilFields(value any) bool {
 	}
 
 	return false
+}
+
+func getEventsTopics(c *Context, w http.ResponseWriter, r *http.Request) {
+	// parameter withSchema is to be parsed
+	withSchema, _ := strconv.ParseBool(r.FormValue("withSchema"))
+
+	topics := c.App.EventBroker().EventTypes()
+
+	if !withSchema {
+		for i := range topics {
+			topics[i].Schema = ""
+		}
+	}
+
+	// calls broker to get a list of topics
+	if err := json.NewEncoder(w).Encode(topics); err != nil {
+		c.Err = model.NewAppError("getTopics", "api.marshal_error", nil, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
+
+func getEventsTopic(c *Context, w http.ResponseWriter, r *http.Request) {
+	c.RequireTopicId()
+	if c.Err != nil {
+		return
+	}
+
+	// parameter withSchema is to be parsed
+	// parameter topicId
+	// parameter withSchema is to be parsed
+	withSchema, _ := strconv.ParseBool(r.FormValue("withSchema"))
+	topicId := c.Params.TopicId
+
+	topics := c.App.EventBroker().EventTypes()
+
+	topic := getTopicByIdFromTopics(topicId, topics)
+
+	if !withSchema {
+		topic.Schema = ""
+	}
+
+	// calls broker to get details of topic with id=topicId
+	if err := json.NewEncoder(w).Encode(topic); err != nil {
+		c.Err = model.NewAppError("getTopics", "api.marshal_error", nil, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
+
+func getEventsSchema(c *Context, w http.ResponseWriter, r *http.Request) {
+	// calls broker to get a list of topics
+	// parameter topicId
+	topicId := c.Params.TopicId
+
+	topics := c.App.EventBroker().EventTypes()
+
+	topic := getTopicByIdFromTopics(topicId, topics)
+
+	if _, err := w.Write([]byte(topic.Schema)); err != nil {
+		c.Err = model.NewAppError("getTopics", "api.marshal_error", nil, err.Error(), http.StatusInternalServerError)
+		return
+	}
+}
+
+func getTopicByIdFromTopics(topicId string, topics []eventbus.EventType) eventbus.EventType {
+	var topic eventbus.EventType
+	for _, t := range topics {
+		if t.Topic == topicId {
+			topic = t
+			break
+		}
+	}
+	return topic
 }
